requirements.txt
Django==1.6.1
PyRSS2Gen==1.0.0
South==0.7.6
argparse==1.2.1
bdist-mpkg==0.4.4
colorama==0.2.4
geopy==0.95
psycopg2==2.4.6
py==1.4.12
pyOpenSSL==0.12
pytest==2.3.4
python-dateutil==1.5
python-gcm==0.1.4
pytz==2012h
requests==0.14.2
wsgiref==0.1.2
xattr==0.6.1
zope.interface==3.5.1
PIL==1.1.7
django-widget-tweaks==1.3
djangorestframework==2.3.1
